chore: rename role route component
