+++ src/components/routing/EnhancedRoleProtectedRoute.tsx
